chore(rename): rename install-dependencies to install
